Move PhaseTimeLeft to components folder
